pyproject.toml
[tool.poetry]
name = "terra-mars-api"
version = "0.2.0"
description = "A Web API for retrieving players' data about their games."
license = "MIT"
authors = ["Rokas Bagdonas <rokius.bagdonas@gmail.com>"]
repository = "https://github.com/RokasBagdonas/terra-mars-api"
keywords = ["terraforming-mars", "api", "Django", "Python", "Web"]

[tool.poetry.dependencies]
python = "^3.8"

django = "^3.0.7"
psycopg2-binary = "^2.8"
djangorestframework = "^3.11.0"
daphne = "^2.5.0"
whitenoise = "^5.1.0"
ipython = "^7.16.1"
django-extensions = "^3.0.3"
django-model-utils = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
pytest-django = "^3.9.0"
flake8 = "^3.8.3"
flake8-docstrings = "^1.5.0"
flake8-pytest-style = "^1.2.1"
pylint = "^2.5.3"
flake8-comprehensions = "^3.2.3"
pytest-randomly = "^3.4.1"
pytest-cov = "^2.10.0"
pytest-factoryboy = "^2.0.3"

[tool.black]
line-length = 90

[tool.coverage.run]
branch = true
dynamic_context = "test_function"

[tool.coverage.report]
precision = 2
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING",
    "raise AssertionError",
    "raise NotImplementedError"
]

[tool.coverage.html]
show_contexts = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

